requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
pyodbc==4.0.34
python-dotenv==0.20.0
SQLAlchemy==1.4.40
Werkzeug==2.2.2
zipp==3.8.1
